+++ app/Http/Controllers/Api/TelegramCommands/PublishCommand.php
@@ -34,6 +34,6 @@ class PublishCommand extends Command
         // the user/chat id who triggered this command.
         // `replyWith<Message|Photo|Audio|Video|Voice|Document|Sticker|Location|ChatAction>()` all the available methods are dynamically
         // handled when you replace `send<Method>` with `replyWith` and use the same parameters - except chat_id does NOT need to be included in the array.
-        $this->replyWithMessage(['text' => "Que tal " . $this->getUpdate()->message->from->username . "!!.\n\nPublicar anuncios es GRATIS, solo tienes que acceder aqui:\n\nhttps://www.bachecubano.com/add"]);
+        @$this->replyWithMessage(['text' => "Que tal " . $this->getUpdate()->message->from->username . "!!.\n\nPublicar anuncios es GRATIS, solo tienes que acceder aqui:\n\nhttps://www.bachecubano.com/add"]);
     }
 }